update tokenCan Sanctum

+++ app/Http/Controllers/ProductController.php
@@ -26,13 +26,22 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required|min:5',
-            'slug' => 'required',
-            'price' => 'required'
-        ]);
+        // เช็คสิทธิ์ (role) ว่าเป็น admin (1) 
+        $user = auth()->user();
 
-        return Product::create($request->all());
+        if($user->tokenCan("1")){
+            $request->validate([
+                'name' => 'required|min:5',
+                'slug' => 'required',
+                'price' => 'required'
+            ]);
+
+            return Product::create($request->all());
+        }else{
+            return [
+                'status' => 'Permission denied to create'
+            ];
+        }
     }
 
     /**
@@ -55,10 +64,19 @@ class ProductController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $product = Product::find($id);
-        $product->update($request->all());
+        // เช็คสิทธิ์ (role) ว่าเป็น admin (1) 
+        $user = auth()->user();
 
-        return $product;
+        if($user->tokenCan("1")){
+            $product = Product::find($id);
+            $product->update($request->all());
+
+            return $product;
+        }else{
+            return [
+                'status' => 'Permission denied to create'
+            ];
+        }
     }
 
     /**
@@ -69,6 +87,17 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
-        return Product::destroy($id);
+        
+        // เช็คสิทธิ์ (role) ว่าเป็น admin (1) 
+        $user = auth()->user();
+
+        if($user->tokenCan("1")){
+            return Product::destroy($id);
+        }else{
+            return [
+                'status' => 'Permission denied to create'
+            ];
+        }
+
     }
 }
